Submit code optimization
