+++ a1_msgs/msg/Pose.msg
@@ -1,0 +1,4 @@
+float32 yaw
+float32 pitch
+float32 roll
+float32 bodyheight
