+++ examples/src/client_get_state.cpp
@@ -2,7 +2,8 @@
 #include <chrono>
 #include <unistd.h>
 #include "rclcpp/rclcpp.hpp"
-#include "comm/comm_ros.h"
+#include "a1_comm/A1_comm.h"
+
 
 using namespace std::chrono_literals;
